-Modify streamCreate()

+++ kuda_runtime_api.h
@@ -48,7 +48,7 @@ extern "C" {
 	//6.4 Stream Management
 	JNIEXPORT jint JNICALL Java_kuda_RuntimeAPI_ctxResetPersistingL2Cache(JNIEnv* env, jobject obj);
 
-	JNIEXPORT jint JNICALL Java_kuda_RuntimeAPI_streamCreate(JNIEnv* env, jobject obj);
+	JNIEXPORT jlong JNICALL Java_kuda_RuntimeAPI_streamCreate(JNIEnv* env, jobject obj);
 
 	//6.5 Event ManageMent
 	JNIEXPORT jint JNICALL Java_kuda_RuntimeAPI_eventCreate(JNIEnv* env, jobject obj, jobject event);
